pyproject.toml
[project]
name = "auditpal"
version = "0.1.0"
description = "AI-powered document assistant for accountants, built on NotebookLM"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "AuditPal Team" }
]
keywords = ["notebooklm", "accounting", "ai", "streamlit", "document-analysis"]

dependencies = [
    "streamlit>=1.32.0",
    "notebooklm-py[browser]>=0.3.0",
    "fpdf2>=2.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
auditpal = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
